Changed conversion from temperature to ADC values to use decimal numbers for more precision.

+++ main.c
@@ -123,7 +123,7 @@ int main (void) {
 			if (EQ_CMD(cmd, ":SET")) {
 				if (EQ_SUBCMD(cmd, ":SET", ":SETPOINT")) {
 					int16_t setpoint = atoi(SUBSTR(cmd, ":SET:SETPOINT "));
-					setpoint = temperature_to_ADC_Pt1000((int8_t)setpoint);
+					setpoint = temperature_to_ADC_Pt1000(setpoint);
 					PID_controller_settings[0].sumError = 0;
 					PID_controller_settings[1].sumError = 0;
 					PID_controller_setpoint = setpoint;
